+++ pages/admin_log_gruppenkasse.inc.php
@@ -8,7 +8,7 @@
 
 $wer = getOrDefault($_GET, 'wer');
 $wen = getOrDefault($_GET, 'wen');
-$gruppe = getOrDefault($_GET, 'gruppe', 0);
+$gruppe = getOrDefault($_GET, 'gruppe', -1);
 $offset = getOrDefault($_GET, 'o', 0);
 ?>
 <div id="SeitenUeberschrift">
@@ -41,11 +41,12 @@ $offset = getOrDefault($_GET, 'o', 0);
         <th>Wohin</th>
     </tr>
     <?php
-    $filter_wer = empty($wer) ? "%" : $wer;
-    $filter_wen = empty($wen) ? "%" : $wen;
-    $entriesCount = Database::getInstance()->getAdminGroupTreasuryLogCount($filter_wer, $filter_wen, $gruppe);
+    $filter_wer = empty($wer) ? null : $wer;
+    $filter_wen = empty($wen) ? null : $wen;
+    $filter_gruppe = $gruppe === -1 ? null : $gruppe;
+    $entriesCount = Database::getInstance()->getAdminGroupTreasuryLogCount($filter_wer, $filter_wen, $filter_gruppe);
     $offset = verifyOffset($offset, $entriesCount, Config::getInt(Config::SECTION_BASE, 'admin_log_page_size'));
-    $entries = Database::getInstance()->getAdminGroupTreasuryLogEntries($filter_wer, $filter_wen, $gruppe, $offset, Config::getInt(Config::SECTION_BASE, 'admin_log_page_size'));
+    $entries = Database::getInstance()->getAdminGroupTreasuryLogEntries($filter_wer, $filter_wen, $filter_gruppe, $offset, Config::getInt(Config::SECTION_BASE, 'admin_log_page_size'));
 
     for ($i = 0; $i < count($entries); $i++) {
         $row = $entries[$i];
@@ -65,7 +66,10 @@ $offset = getOrDefault($_GET, 'o', 0);
     }
     ?>
 </table>
-<?= createPaginationTable('/?p=admin_log_gruppenkasse&amp;wer=' . escapeForOutput($wer) . '&amp;wen=' . escapeForOutput($wen) . '&amp;gruppe=' . $gruppe, $offset, $entriesCount, Config::getInt(Config::SECTION_BASE, 'admin_log_page_size')); ?>
+<?= createPaginationTable('/?p=admin_log_gruppenkasse&amp;wer=' . escapeForOutput($wer)
+    . '&amp;wen=' . escapeForOutput($wen)
+    . '&amp;gruppe=' . $gruppe
+    , $offset, $entriesCount, Config::getInt(Config::SECTION_BASE, 'admin_log_page_size')); ?>
 
 <div>
     <a href="/?p=admin">&lt;&lt; Zurück</a>